+++ src/components/JqueryComponent.js
